Refactored terrain methods so it inherits model loading from world objects
