+++ Spark/Classes/WorldState.swift
@@ -8,6 +8,13 @@
 import Foundation
 import SceneKit
 
+// how to add terrain to scene
+//terrainScene?.rootNode.enumerateChildNodes({ node, stop in
+//    if let name = node.name, name.contains(nodeName){
+//        modelCopy = referenceNode!.clone()
+//         node name referes to parent node of terrain
+//        node.addChildNode(modelCopy!)
+//    }
 class WorldState{
     var terrain:Terrain!
